# api/src/utils/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Verbify - OpenAPI 3.0
  description: |-
    Verbify API endpoints - Server based on the OpenAPI 3.0 specification.  
  version: 1.0.11
servers:
  - url: http://localhost:8080/api/
...
paths:
...
  /users:
    post:
      tags:
        - users
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
          application/xml:
            schema:
              $ref: '#/components/schemas/Users'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
            application/xml:
              schema:
                $ref: '#/components/schemas/Users'
...
components:
  schemas:
...
    Users:
      type: object
      properties:
        id:
          type: string
          format: objectId
          example: '6427c38d08b86ce760a1fa47'
        displayName:
          type: string
          example: theUser_99
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: john@email.com
        passwordHash:
          type: string
          example: '12345'
        isPremium:
          type: boolean
          example: true
        roles:
          type: array
          xml:
            name: roles
            wrapped: false
          items:
            $ref: '#/components/schemas/Roles'
      xml:
        name: users
    Roles:
      type: object
      properties:
        name:
          type: string
          example: 'admin'
      xml:
        name: roles
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header